requirements.txt
aiodine==1.2.7
astroid==2.2.5
bocadillo==0.18.2
bocadillo-cli==0.2.2
certifi==2019.6.16
chardet==3.0.4
Click==7.0
h11==0.8.1
httptools==0.0.13
idna==2.8
isort==4.3.21
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-engineio==3.9.3
python-multipart==0.0.5
python-socketio==4.3.1
requests==2.22.0
six==1.12.0
starlette==0.12.9
typed-ast==1.4.0
typesystem==0.2.4
urllib3==1.25.3
uvicorn==0.8.6
uvloop==0.12.2
websockets==7.0
whitenoise==4.1.3
wrapt==1.11.2
